currencyDataScreen-done-styled.
