+++ app/src/main/java/com/example/ratemate/data/api/ExchangeRatesApi.kt
@@ -1,6 +1,5 @@
 package com.example.ratemate.data.api
 
-import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
 
@@ -8,4 +7,13 @@ interface ExchangeRatesApi {
 
     @GET("latest?access_key=baaeaff430616fe496d196d12c6587a0")
     suspend fun getExchangeRates(): ExchangeRatesResponse
+
+    @GET("historical")
+    suspend fun getHistoricalRates(
+        @Query("access_key") accessKey: String,
+        @Query("start_date") startDate: String,
+        @Query("end_date") endDate: String,
+        @Query("symbols") symbols: String
+    ): HistoricalRatesResponse
 }
+
