+++ app/src/main/java/com/example/ratemate/data/api/ExchangeRatesViewModel.kt
@@ -12,6 +12,7 @@ import com.example.ratemate.data.database.CurrencyRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 
 class ExchangeRatesViewModel(application: Application) : AndroidViewModel(application) {
@@ -21,8 +22,13 @@ class ExchangeRatesViewModel(application: Application) : AndroidViewModel(applic
 
     // Expose all currencies as a Flow
     val currencies: Flow<List<CurrencyEntity>> = repository.getAllCurrencies()
-
     val favorites: Flow<List<CurrencyEntity>> = repository.getFavoritedCurrencies()
+
+    private val _selectedCurrencyRate = MutableStateFlow<Double?>(null)
+    val selectedCurrencyRate: StateFlow<Double?> = _selectedCurrencyRate.asStateFlow()
+
+    private val _historicalRates = MutableStateFlow<Map<String, Double>?>(null)
+    val historicalRates: StateFlow<Map<String, Double>?> = _historicalRates.asStateFlow()
 
     fun toggleFavoriteStatus(currencyCode: String, isFavorited: Boolean) {
         viewModelScope.launch {
@@ -46,7 +52,44 @@ class ExchangeRatesViewModel(application: Application) : AndroidViewModel(applic
             }
         }
     }
+
+    fun fetchExchangeRateForCurrency(currencyCode: String) {
+        viewModelScope.launch {
+            try {
+                val response = ApiClient.apiService.getExchangeRates()
+                val rate = response.rates[currencyCode]
+                _selectedCurrencyRate.value = rate
+            } catch (e: Exception) {
+                // Handle error
+                println("Error fetching exchange rate: $e")
+                e.printStackTrace()
+            }
+        }
+    }
+
+
+    fun fetchHistoricalRates(currencyCode: String, startDate: String, endDate: String) {
+        viewModelScope.launch {
+            try {
+                val response = ApiClient.apiService.getHistoricalRates(
+                    accessKey = "baaeaff430616fe496d196d12c6587a0",
+                    startDate = startDate,
+                    endDate = endDate,
+                    symbols = currencyCode
+                )
+                println("Historical rates response: $response")
+                _historicalRates.value = response.rates[currencyCode]
+            } catch (e: Exception) {
+                println("Error fetching historical rates: $e")
+                e.printStackTrace()
+            }
+        }
+    }
+
+
 }
+
+
 
 
 class ExchangeRatesViewModelFactory(
