+++ app/src/main/java/com/example/ratemate/data/api/HistoricalRatesResponse.kt
@@ -1,0 +1,8 @@
+package com.example.ratemate.data.api
+
+data class HistoricalRatesResponse(
+    val success: Boolean,
+    val rates: Map<String, Map<String, Double>>,
+    val start_date: String,
+    val end_date: String
+)
