+++ app/src/main/java/com/example/ratemate/ui/screens/ConvertScreen.kt
@@ -6,9 +6,12 @@ import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.navigation.NavHostController
+import com.example.ratemate.data.api.ExchangeRatesViewModel
 
 @Composable
-fun ConvertScreen() {
+fun ConvertScreen(navController: NavHostController, viewModel: ExchangeRatesViewModel = viewModel() ) {
     Box(
         modifier = Modifier.fillMaxSize(),
         contentAlignment = Alignment.Center
