+++ app/src/main/java/com/example/ratemate/ui/screens/CurrencyScreen.kt
@@ -1,11 +1,6 @@
 package com.example.ratemate.ui.screens
 
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
 
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.runtime.LaunchedEffect
@@ -14,10 +9,14 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.ratemate.ui.components.CurrencyCard
 import com.example.ratemate.data.api.ExchangeRatesViewModel
 import androidx.compose.ui.res.painterResource
+import androidx.navigation.NavHostController
 import com.example.ratemate.R
 
 @Composable
-fun CurrencyScreen(viewModel: ExchangeRatesViewModel = viewModel()) {
+fun CurrencyScreen(
+    navController: NavHostController,
+    viewModel: ExchangeRatesViewModel = viewModel()
+) {
     // Trigger data fetching when the screen is first displayed
     LaunchedEffect(Unit) {
         viewModel.fetchAndSaveExchangeRates()
